move EventPage to EventListItems
